+++ test/test_helper.rb
@@ -24,8 +24,9 @@ class Minitest::Test
   end
 
   def teardown
-    ScoutApm::Agent.instance.shutdown
+    ScoutApm::Agent.instance.shutdown if ScoutApm::Agent.instance.started?
     File.delete(DATA_FILE_PATH) if File.exist?(DATA_FILE_PATH)
+    ScoutApm::Agent.class_variable_set("@@instance",nil)
   end
 
   def set_rack_env(env)