+++ test/unit/config_test.rb
@@ -26,5 +26,18 @@ class ConfigTest < Minitest::Test
     assert_equal "debug", conf.value('log_level')
     assert_equal "APM Test Conf (Production)", conf.value('name')
   end
+
+  # Ensure a bad config file doesn't cause an exception when accessing a config value
+  def test_config_value_with_bad_config
+    ScoutApm::Agent.class_variable_set("@@instance",nil) # need to reset this so we're ready with a new agent
+    agent = ScoutApm::Agent.instance(:config_path => File.expand_path("../../data/bad_config.yml", __FILE__))
+    no_error = true
+    begin
+      agent.config.value('monitor')
+    rescue Exception => e
+      no_error = false
+    end
+    assert no_error, "Error fetching config value w/bad config file: #{e.message if e}"
+  end
 end