+++ CHANGELOG.markdown
@@ -2,7 +2,7 @@
 
 * Summarizing middleware instrumentation into a single metric for lower overhead.
 * If monitoring isn't enabled:
-  * Don't install ScoutApm::Middleware
+  * In ScoutApm::Middleware, don't start the agent
   * Don't start background worker
 
 # 1.2.10